authentication/requirements.txt
# Framework Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

pydantic>=2.6.0
pydantic-settings>=2.1.0

python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
bcrypt==4.0.1

httpx>=0.26.0
tenacity>=8.2.0

structlog>=24.1.0